requirements.txt
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.1.5
django-filter==2.0.0
django-phonenumber-field==2.0.1
django-registration==3.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djangorestframework-simplejwt==4.3.0
idna==2.7
Naked==0.1.31
phonenumbers==8.9.16
Pillow==5.3.0
psycopg2==2.7.5
PyJWT==1.7.1
pytz==2018.5
PyYAML==3.13
requests==2.20.0
status==0.2.5
urllib3==1.24
